+++ src/main/java/com/luizfd/MDORM/entities/Participante.java
@@ -17,7 +17,7 @@ public class Participante {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	@Column(unique = true)
 	private String email;
@@ -29,18 +29,18 @@ public class Participante {
 		
 	}
 
-	public Participante(Long id, String nome, String email) {
+	public Participante(Integer id, String nome, String email) {
 		super();
 		this.id = id;
 		this.nome = nome;
 		this.email = email;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
